pyproject.toml
[tool.poetry]
name = "datacheck"
version = "0.1.0"
description = "Lightweight data quality validation CLI tool"
authors = ["datacheck <hello@datacheck.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/yash-chauhan-dev/datacheck"
repository = "https://github.com/yash-chauhan-dev/datacheck"
keywords = ["data-quality", "validation", "cli", "data-engineering"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.12.0"
pandas = "^2.0.0"
pyyaml = "^6.0"
rich = "^13.0.0"
pyarrow = "^22.0.0"
duckdb = "0.8.1"
click = "~8.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
mypy = "^1.5.0"
ruff = "^0.1.0"
pre-commit = "^3.5.0"
types-pyyaml = "^6.0.0"
pandas-stubs = "^2.0.0"

[tool.poetry.scripts]
datacheck = "datacheck.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.isort]
known-first-party = ["datacheck"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = "duckdb.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=datacheck",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=json",
    "--cov-fail-under=70",
    "-ra",  # Show all test summary info
    "--tb=short",  # Shorter traceback format
]
markers = [
    "unit: Unit tests that don't require external dependencies",
    "integration: Integration tests that test multiple components",
    "slow: Tests that take longer to run",
    "cli: CLI-specific tests",
    "loader: Data loader tests",
    "rules: Validation rules tests",
    "config: Configuration tests",
    "engine: Validation engine tests",
    "output: Output formatting tests",
]
filterwarnings = [
    "ignore::DeprecationWarning:typer.*",  # Ignore typer deprecation warnings
    "ignore::DeprecationWarning:click.*",  # Ignore click deprecation warnings
    "ignore::pandas.errors.SettingWithCopyWarning",  # Ignore pandas copy warnings
]

[tool.coverage.run]
source = ["datacheck"]
omit = ["tests/*", "**/__main__.py"]
branch = true  # Enable branch coverage

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "def __str__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.json]
output = "coverage.json"
